+++ testcafe/page-object/editor/container-contents.ts
@@ -36,4 +36,8 @@ export namespace containerContents {
     export function getFirstListItem() {
         return new NodeRow(Selector('mesh-node-row').nth(0));
     }
+
+    export function getNumberOfItems() {
+        return Selector('mesh-node-row').count;
+    }
 }